requirements.txt
python-telegram-bot>=21.6,<23
python-dotenv>=1.0.0
httpx>=0.27,<0.29
cryptography>=42.0.0
requests>=2.31.0
urllib3>=2.0.0
PySocks>=1.7.1
